move unit tests around
